% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_markerslist.R
\name{make_markerslist}
\alias{make_markerslist}
\title{Helper function to demonstrate how to
make a markerslist compatible with \code{fit_metagene_scores()} function.}
\usage{
make_markerslist(
  index_marker,
  predictors,
  signs = NULL,
  use_offclass_markers_as_negative_predictors = TRUE,
  return_as_datatable = FALSE
)
}
\arguments{
\item{index_marker}{}

\item{predictors}{}

\item{signs}{an optional list of signs for each cell type class, i.e., c(1,1, -1) corresponding to the expected association direction
of the predictors with the cell type class.
If left empty (NULL is default),
all predictors are assumed to be positively associated
with the cell type class.}

\item{use_offclass_markers_as_negative_predictors}{if TRUE, predictors of a cell type class are added as negative predictors
of other celltype classes
(unless they are specified as positive predictors in both classes).}
}
\value{
a 'markerslist' list object which can be passed to \code{fit_metagene_scores}
}
\description{
Each element of the returned 'markerslist' corresponds to a cell type class, and
Typically, one would want the cell type classes in a 'markerslist' to be mutually exclusive.
See examples for more details.
}
\examples{

markerslist <-  
  make_markerslist(
     index_marker = list(bcell = c("CD19")
                          ,plasma = c("IGKC")
                          ,myeloid = "LYZ"
                          ,tcell = "CD3E"
                          )
     ,predictors = list(tcell = c("CD3D", "CD3E", "CD3G", "CD2", "CD4", "CD8A"  
                                  ,"CD8B", "IL2" , "IL4", "IL17A", "IFNG",   "PRF1"  
                                  ,"GZMA", "GZMB","FOXP3"
                                  )
                        ,bcell = c("CD19", "MS4A1", "CD79A", "CD79B", "PAX5"
                                     ,"CD22", "IGKC", "TCL1A", "CD40"
                                   )
                        ,myeloid = c("CD163", "CD68","CSF1R","CD4", "ITGAM", "ITGAX","LYZ"     
                                     ,"FCGR3A", "MRC1", "CD14" ,"HLA-DRA", "HLA-DRB1"
                                     )
                        ,plasma = c("IGHM", "IGHG1", "IGHA1", "IGKC"
                                    ,"IGHA1", "IGHD", "CD38")
                        )
                        
   )
  
}
