% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{totalcount_norm}
\alias{totalcount_norm}
\title{Normalize cell expression in a raw counts matrix by their totalcounts}
\usage{
totalcount_norm(counts_matrix, tc = NULL)
}
\arguments{
\item{counts_matrix}{a cells x genes matrix of raw counts to be normalized}

\item{tc}{optional vector of totalcounts.  Useful if providing a counts_matrix based on a subset of genes.
If not provided, totalcounts per cells is taken to be the rowSums of the counts matrix.}
}
\description{
Normalize cell expression in a raw counts matrix by their totalcounts
}
